Feat #93 - utils/categoryFilter 파일 생성 및 IProduct category -> categoryId 로 변경
