+++ src/components/UI/organisms/DetailBox/DetailBox.stories.tsx
@@ -31,7 +31,7 @@ MyProduct.args = {
   },
   productDetail: {
     title: '아이폰 삽니다',
-    category: '디지털기기',
+    categoryId: 1,
     description: '원가 18,900갤럭시로 갈아타면서 판매해요~ 한번도 사용안한 새상품입니다',
     created_at: new Date(),
     view: 256,
@@ -60,7 +60,7 @@ Not_MyProduct.args = {
   },
   productDetail: {
     title: '아이폰 삽니다',
-    category: '디지털기기',
+    categoryId: 1,
     description: '원가 18,900갤럭시로 갈아타면서 판매해요~ 한번도 사용안한 새상품입니다',
     created_at: new Date(),
     view: 256,
